+++ FinanceCompanyBackend/src/main/java/com/java/dataaccess/contracts/DocumentContract.java
@@ -6,7 +6,7 @@ import com.java.entities.Documents;
 
 public interface DocumentContract {
 	boolean addDocument(Documents docs) throws Exception ;
-	boolean removeDocuments(int application_number) throws Exception ;
+	boolean removeDocuments(long application_number) throws Exception ;
 	List<Documents> getAllDocuments() throws Exception;
-	Documents getDocumentById(int application_number) throws Exception;
+	Documents getDocumentById(long application_number) throws Exception;
 }
