+++ FinanceCompanyBackend/src/main/java/com/java/dataaccess/contracts/LoanApplicationContract.java
@@ -8,8 +8,8 @@ import com.java.entities.LoanApplication;
 public interface LoanApplicationContract {
 	boolean addApplication(LoanApplication application) throws Exception ;
 	List<FullApplication> getAllApplications() throws Exception ;
-	List<FullApplication> getApplicationsById(int customer_id) throws Exception ;
-	boolean removeApplication(int application_number) throws Exception ;
-	boolean approve(int application_number) throws Exception ;
-	boolean reject(int application_number) throws Exception ;
+	List<FullApplication> getApplicationsById(long customer_id) throws Exception ;
+	boolean removeApplication(long application_number) throws Exception ;
+	boolean approve(long application_number) throws Exception ;
+	boolean reject(long application_number) throws Exception ;
 }
